PAE-1590: Move the committed row-state collection to the waste-records aggregate
The committed row-state collection is the canonical committed state of each
waste record, so it belongs to the waste-records aggregate rather than
waste-balances. Relocate the collection, its repository and the current-head
read model to a record-owned home.

The balance stream stays the producer: it emits into the collection through an
injected repository it no longer constructs. The record side now owns and
registers that repository; the balance plugin consumes it from the server.
Classification stays co-stored on each row-state document. No behaviour change.
