+++ src/repositories/waste-records/committed-row-states/contract/findBySummaryLogId.contract.js
@@ -1,9 +1,6 @@
 import { describe, beforeEach, expect } from 'vitest'
 
-import {
-  buildRowStateEntry,
-  DEFAULT_PARTITION
-} from '../row-states-test-data.js'
+import { buildRowStateEntry, DEFAULT_PARTITION } from '../test-data.js'
 
 export const testFindBySummaryLogIdBehaviour = (it) => {
   describe('findBySummaryLogId', () => {
