+++ src/repositories/waste-records/committed-row-states/contract/upsertRowStates.contract.js
@@ -3,10 +3,7 @@ import { describe, beforeEach, expect } from 'vitest'
 import { ROW_OUTCOME } from '#domain/summary-logs/table-schemas/validation-pipeline.js'
 import { WASTE_RECORD_TYPE } from '#domain/waste-records/model.js'
 
-import {
-  buildRowStateEntry,
-  DEFAULT_PARTITION
-} from '../row-states-test-data.js'
+import { buildRowStateEntry, DEFAULT_PARTITION } from '../test-data.js'
 
 const excludedClassification = {
   outcome: ROW_OUTCOME.EXCLUDED,
