+++ src/repositories/waste-records/committed-row-states/inmemory.plugin.js
@@ -1,0 +1,17 @@
+import { createInMemoryRowStateRepository } from './inmemory.js'
+import { registerRepository } from '#plugins/register-repository.js'
+
+export function createInMemoryCommittedRowStatesRepositoryPlugin() {
+  const repository = createInMemoryRowStateRepository()()
+
+  return {
+    name: 'committedRowStatesRepository',
+    register: (server) => {
+      registerRepository(
+        server,
+        'committedRowStatesRepository',
+        () => repository
+      )
+    }
+  }
+}
