+++ src/repositories/waste-records/committed-row-states/mongodb.plugin.js
@@ -1,0 +1,15 @@
+import { createMongoRowStateRepository } from './mongodb.js'
+import { registerRepository } from '#plugins/register-repository.js'
+
+export const mongoCommittedRowStatesRepositoryPlugin = {
+  name: 'committedRowStatesRepository',
+  version: '1.0.0',
+  dependencies: ['mongodb'],
+
+  register: async (server) => {
+    const factory = await createMongoRowStateRepository(server.db)
+    const repository = factory()
+
+    registerRepository(server, 'committedRowStatesRepository', () => repository)
+  }
+}
