+++ src/test/create-test-server.js
@@ -35,6 +35,7 @@ import { createInMemorySummaryLogsRepositoryPlugin } from '#repositories/summary
 import { createInMemorySystemLogsRepositoryPlugin } from '#repositories/system-logs/inmemory.plugin.js'
 import { createInMemoryWasteBalancesRepositoryPlugin } from '#waste-balances/repository/inmemory.plugin.js'
 import { createInMemoryWasteRecordsRepositoryPlugin } from '#repositories/waste-records/inmemory.plugin.js'
+import { createInMemoryCommittedRowStatesRepositoryPlugin } from '#repositories/waste-records/committed-row-states/inmemory.plugin.js'
 
 /** @import { Lifecycle } from '@hapi/hapi' */
 /** @import { LogMethod } from '#common/helpers/logging/logger.js' */
@@ -102,6 +103,10 @@ const repositoryConfigs = [
   {
     name: 'wasteRecordsRepository',
     createDefault: createInMemoryWasteRecordsRepositoryPlugin
+  },
+  {
+    name: 'committedRowStatesRepository',
+    createDefault: createInMemoryCommittedRowStatesRepositoryPlugin
   },
   {
     name: 'wasteBalancesRepository',
