+++ src/waste-balances/repository/mongodb.test.js
@@ -9,7 +9,7 @@ import {
 import {
   createMongoRowStateRepository,
   WASTE_BALANCE_ROW_STATES_COLLECTION_NAME
-} from './row-states-mongodb.js'
+} from '#repositories/waste-records/committed-row-states/mongodb.js'
 import { buildStreamEvent } from './stream-test-data.js'
 import { testWasteBalancesRepositoryContract } from './port.contract.js'
 
@@ -25,7 +25,7 @@ const DATABASE_NAME = 'epr-backend'
  * @typedef {object} WasteBalancesRepoFixtures
  * @property {import('mongodb').MongoClient} mongoClient
  * @property {import('./stream-port.js').WasteBalanceStreamRepository} streamRepository
- * @property {import('./row-states-port.js').RowStateRepository} rowStateRepository
+ * @property {import('#repositories/waste-records/committed-row-states/port.js').RowStateRepository} rowStateRepository
  * @property {import('./port.js').WasteBalancesRepositoryFactory} wasteBalancesRepository
  * @property {(event: object) => Promise<void>} seedBalance
  */
